iqranow/server/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.31
Werkzeug==3.0.3
python-dotenv==1.0.1
google-cloud-speech==2.25.0
pydub==0.25.1
itsdangerous==2.2.0
gunicorn==21.2.0
